auhsfiul

+++ main.cpp
@@ -61,7 +61,7 @@ int main(int argc, char *argv[])
     //cout<<asdfgh<<asdfgh<<qwe<<asdfgh<<qwe<<endl;
     //219 -> 178 -> 177 -> 176 -> 43
     
-   short asdf = 24;
+   long long asdf = 24;
    cout<<sizeof(asdf)<<endl;
    
    char buffer[64];
